Adjusted views

+++ db/migrate/20160404190209_create_order_items.rb
@@ -4,7 +4,7 @@ class CreateOrderItems < ActiveRecord::Migration
       t.string :purchaser_name, null: false
       t.string :item_description, null: false
       t.decimal :item_price, null: false
-      t.decimal :purchase_count, null: false
+      t.integer :purchase_count, null: false
       t.string :merchant_address, null: false
       t.string :merchant_name, null: false
       t.decimal :total_price
